Refactor secure and monitor into it's own files
_client.py has grown too big. Moving SdcMonitor and SdcSecure into
separated files.
