+++ sdcclient/__init__.py
@@ -1,3 +1,3 @@
-from sdcclient._client import SdcClient
-from sdcclient._client import SdMonitorClient
-from sdcclient._client import SdSecureClient
+from sdcclient._monitor import SdcClient
+from sdcclient._monitor import SdMonitorClient
+from sdcclient._secure import SdSecureClient
